Adds the confusion matrix to the boosting.py file.

+++ code/BoostingWithGridSearch.py
@@ -13,7 +13,7 @@ spark = SparkSession.builder.appName('Boosted MLPs').getOrCreate()
 
 schema = StructType([
     StructField('station', StringType(), False),
-    StructField('dateofyear', IntegerType(), False),
+    StructField('dateofyear', FloatType(), False),
     StructField('latitude', FloatType(), False),
     StructField('longitude', FloatType(), False),
     StructField('elevation', FloatType(), False),
